cfile_set: don't reset BloomFileReader on error
There is a race between UpdateCompactionStats and the error path of
CFileSet::FindRow(). When FindRow() fails to check its BloomFileReader,
it clears it, but the reader is necessary to determine the rowset size.
Both access the bloom reader without locks.

The reset is an optimization to allow the CFileSet to be scanned
post-error without having to read the bloom file. This isn't compelling:
the failure of a bloom reader doesn't affect the correctness of a scan,
and if the error is a permanent error (e.g. a disk failure), it would be
desirable to eventually fail and replace the tablet; if the error is
transient, then re-reading the bloom filter could yield successful
scans. In either case, clearing the reader doesn't seem necessary.

This patch removes the BloomFileReader reset on error. A couple of
places previously checked the presence of the bloom reader because of
the reset; this is no longer necessary. Also includes some updates to
style in a few related places.

Change-Id: I7435da448a1a33ca5216923311ab920f7a859806
Reviewed-on: http://gerrit.cloudera.org:8080/8440
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>