[build] fix OPENSSL_ROOT_DIR override on RH/CentOS
Due to an extra level of indirection in the condition for one if()
expression, the OpenSSL location was not overridden on RedHat/CentOS
machines where the OpenSSL workaround build existed in
$KUDU_ROOT/thirdparty/installed/openssl-el6-workaround,
even if the OPENSSL_ROOT_DIR was overridden for cmake invocation
(i.e. -DOPENSSL_ROOT_DIR=/path/to/alt-openssl-location was specified).

Removing the level of indirection for the OPENSSL_ROOT_DIR variable
in the expression fixed the issue, which seems to be in sync with
  https://cmake.org/cmake/help/v3.9/command/if.html

At least, with this patch the issue is gone if running cmake 3.9.0.

Change-Id: Ibe69b5d62ef4054e60138a154a85569d92166aff
Reviewed-on: http://gerrit.cloudera.org:8080/8407
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>