[mini-cluster] allow 18-bit PID in LOOPBACK bind mode
This changelist allows to use 18-bit PIDs for binding in LOOPBACK mode.
Prior to this patch, the minicluster could not use PIDs wider
than 16 bits.

While expanding the range of acceptable PIDs, this changelist also
decreases the maximum allowed number of 'indexed servers' from 254
to 62.  As of now, that's enough for our minicluster-based tests.

The incentive to put up this patch was change of the max PID
up to 147456 in /proc/sys/kernel/pid_max after recent restart of
one of our build servers.

Change-Id: I958729d71ed363e98bea99f3d932cc6b0546e130
Reviewed-on: http://gerrit.cloudera.org:8080/8422
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>