cmake: use MAKE_C_IDENTIFIER to generate valid target names
This fixes a build failure coming from the FindThrift cmake module in
the case that the source directory is in a path containing a character
which is not valid for CMake target names. For example, package version
numbers often include the '~' character which isn't allowed in a target
name. This would produce an error like:

09:42:16 CMake Error at cmake_modules/FindThrift.cmake:128 (add_custom_target):
09:42:16   The target name
09:42:16   "-data-jenkins-workspace-generic-package-ubuntu64-16-04-impala-CDH5-Packaging-Kudu-2017-11-06_08-48-50-kudu-1.6.0+cdh5.14.0+0-1.cdh5.14.0.p0.91~xenial-build-release-src-kudu-hms-hive_metastore.thrift"
09:42:16   is reserved or not valid for certain CMake features, such as generator
09:42:16   expressions, and may result in undefined behavior.
09:42:16 Call Stack (most recent call first):
09:42:16   src/kudu/hms/CMakeLists.txt:18 (THRIFT_GENERATE_CPP)

The fix is to use the MAKE_C_IDENTIFIER string function from CMake to ensure
that the generated target name is a valid target name.

Change-Id: I71f145c08d7fbb5627442b6527b3feebc05055bd
Reviewed-on: http://gerrit.cloudera.org:8080/8481
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Dan Burkert <danburkert@apache.org>