KUDU-1454 [part 1]: update propagated timestamp on the driver
Currently, Spark uses multiple clients simultaneously, possibly on
different JVMs, for reads and writes. Each client may have the
propagated timestamp properly updated after each read/write. However,
the KuduClient on the driver does not have the correct propagated
timestamp. This is problematic if the user wants to use READ_AT_SNAPSHOT
ReadMode to get a consistent snapshot.

This patch uses Accumulator in Spark to properly update the propagated
timestamp on the driver through comparing the maximum timestamp on each
executor. It also updates back last propagated timestamp to the
executors for further writes. Test is added accordingly and ran in both
'local' and 'local-cluster' mode.

Change-Id: Id0a078ae8ebaa6a859be75c822879291192c5842
Reviewed-on: http://gerrit.cloudera.org:8080/8552
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>