[RaftPeerPB] introduce attributes and health status
This standalone commit introduces changes to provide necessary interface
for the new 3-4-3 scheme of automated replica replacement.

Attributes and health report information has been added into the
RaftPeerPB message (which represents a tablet replica).

The 'health_report' is a non-persistent optional field which is populated
by leader replica.  It reflects the health of the replica as seen by
the leader.

The 'attrs' field is used as an umbrella for various per-replica
attributes.  The attributes are persisted along with other fields
for RaftPeerPB structure.  Right away, the 'promote' and 'replace'
attributes are introduced: they are necessary for the new scheme
of automated replica replacement:

  * 'promote' is applicable to non-voter replicas only.  When set
     to 'true', it makes the leader replica promoting the non-voter
     replica once it catches up with the leader's WAL.

  * 'replace' is applicable to voter replicas only.  When set, it means
    'replace the replica regardless of its health status'.
    This attribute is needed to provide the capability to manually move
    a replica from its current location using the kudu CLI tool
    (i.e. 'kudu tablet change_config move_replica ...').

Change-Id: Id82f838fd3612ab9bc3f91cac7a840cb9f36ff4c
Reviewed-on: http://gerrit.cloudera.org:8080/8561
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Mike Percy <mpercy@apache.org>