[tools] added verbose mode for ksck
Introduced verbose mode for the ksck tool ('kudu cluster ksck').
The verbose mode is enabled if specifying --verbose flag.  With the
verbose mode, the ksck outputs detailed information on cluster's
metadata, regardless whether the metadata information is found
consistent or not.

Without this patch, the detailed info was output only if the tool
found inconsistency in the cluster.

Change-Id: Ie6688510c90fff5da9742b1496ee805e93dbd4c0
Reviewed-on: http://gerrit.cloudera.org:8080/8596
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>