schema: fast-path for Schema::KeyEquals, etc.
If two schemas are the same object, then KeyEquals can be fast-pathed.
The same goes for various ColumnSchema equality checks.

This hasn't been a hot path in release builds, but I noticed a lot of wasted
CPU in a debug build evaluating Schema::KeyEquals, and my suspicion is that in
a great number of cases the objects are identical. Even in release builds, may
as well avoid the unnecessary cycles.

Change-Id: Ia5fadfd70ea651b9105837535f21a34c6e0af6ae
Reviewed-on: http://gerrit.cloudera.org:8080/8627
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins