disk failure: randomized tserver test
This patch adds a test that tests various paths that perform block IO by
manually applying randomized operations to a tablet and injecting disk
failures a fraction of the time. Eventually, the replica is marked as
failed, completing the test. This patch doesn't test the behavior after
reaching this failed state, but rather ensures this state is reachable
without crashing the server.

Change-Id: I17aca35fc0bf9c7b1b73308def865b0ed48d07fa
Reviewed-on: http://gerrit.cloudera.org:8080/8387
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins