KUDU-1097 (patch 3): Implement promotion of NON_VOTER replicas
This patch implements automatic leader-side promotion of NON_VOTER
replicas that have their "promote" attribute set. This occurs when the
NON_VOTER in question catches up to within a certain number of WAL
entries of the commit index, as tracked by the leader (the default is
100, controlled by --consensus_promotion_max_wal_entries_behind).

This patch also adds some infrastructure to support this functionality:

* New notification callback so the consensus queue can tell
  RaftConsensus when to promote a non-voter.
* Support for including RaftAttrPB attributes in CHANGE_REPLICA_TYPE.
* A simple non-voter promotion test.

More comprehensive end-to-end testing is needed before enabling this
feature by default.

Change-Id: Ife6f59658cb2f38d087d57c244ad811010a91fef
Reviewed-on: http://gerrit.cloudera.org:8080/8633
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins