consensus: Rename ChangeConfigType enum constants
Since we now have attributes in the RaftConfigPeerPB message, it would
be more useful to allow a peer to be modified such that only its
attributes change, but not its member type. In order to provide
meaningful semantics, that means we should change the constant name for
the "modify peer" operation as well. This patch changes
CHANGE_REPLICA_TYPE to MODIFY_PEER.

While we're in there, let's also rename the other constants to more
accurately reflect the terminology we tend to use in this part of the code.

There are no functional or wire-protocol changes in this patch. The
functional changes to go along with this patch will come in a follow-up
patch.

Change-Id: I77b838f1b7ddcf176d1a5146e04fa1fddc1042fa
Reviewed-on: http://gerrit.cloudera.org:8080/8643
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>