tablet_copy-itest: fix flaky TestMetricsResetAfterRevival
The test aimed to assert that metrics would reset after successfully
bootstrapping the tablet. It originally used the condition that the
metadata reached TABLET_DATA_READY to ensure bootstrapping completed,
which may not be the case. E.g. during tablet copies, the data is copied
over in TABLET_DATA_READY state, and then the Tablet is bootstrapped
(and only then are the metrics reset).

This patch updates the test to ensure bootstrapping completes by instead
checking the number of running tablets reaches 1.

I validated this by adding a sleep statement just before OpenTablet() in
the tablet copy code (where the metadata has reached TABLET_DATA_READY
but has yet to bootstrap) and seeing 100/100 failures. With the fix,
this case passes.

Change-Id: Icec1eb2cb8b9569c27210d0342087b1b05169046
Reviewed-on: http://gerrit.cloudera.org:8080/8651
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins