java: revisit maven shade configuration
This is a followup to [1], which reconfigured the Maven Shade Plugin to
apply excludes to all shaded jars, most notably on Guava classes. This
causes issues with kudu-flume-sink, which has a legitimate transitive
dependency on Guava through flume itself.

While investigating this issue I found that the `<include>*:*</include>`
configuration we commonly set for the shade plugin caused more
dependencies to be included than desired, for instance kudu-client-tools
was including hadoop-client, which was in provided scope.

The fix to both of these issues ended up being the same: only set the
exclusions where they should take effect in kudu-client, and be more
explicit about what jars should be included when shading in all other
modules.

I've done my best to make sure that the correct jars are still included
in the shaded jars using `jar -tf`.

[1]: https://github.com/apache/kudu/commit/5a258508f8d560f630512c237711a65cd137c6b3

Change-Id: I8f02a37a5a6683dbbb226a1faf96647a112d7c0d
Reviewed-on: http://gerrit.cloudera.org:8080/8661
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>