java build: Add script to verify output of generated JARs
This script reports any files included in packaging JARs that do not
match a set of patterns in the script. This allows us to define certain
dependencies that we will ship (shaded or otherwise) and, if this is
incorporated into the build process, will immediately alert us as things
change over time.

The script runs automatically as part of the regular CI build.

Change-Id: Ie16bca12dfe2d0267e559c91abe1178d906e3122
Reviewed-on: http://gerrit.cloudera.org:8080/8662
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>