tablet_copy-itest: fix TestMetricsResetAfterRevival pt2
b29c9373471927d3f2dfbdab03a8bd728c73d82f introduced a fix to the flaky
test such that the tablet's state is checked to be RUNNING. It did not
take into account the fact that tablet state metrics are cached, so the
tablet could still be in the process of bootstrapping, while it was
being advertized as a running tablet. As such, when run with 8 stress
threads, this failed 55/100 runs.

This patch addresses this by not caching tablet state. With this, with
stress, this passes 1000/1000 runs.

Change-Id: I06e538cc471d15447a90105a4d6ce25f5a69d2b0
Reviewed-on: http://gerrit.cloudera.org:8080/8665
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>