KUDU-1097: scenario for fallen-behind-the-WAL replica
Added end-to-end test scenario for the case when an existing follower
replica falls behind the WAL GC threshold.  First, the system attempts
to replace it with a non-voter where tablet copying fails.  As soon as
that detected, a new non-voter replica is added and the replacement
succeeds, so the tablet configuration eventually gets rid of failed
replicas and contains exactly 3 healthy replicas.

Change-Id: Ifcaff45c9fa3beb51d99dc53d88138da7dc120af
Reviewed-on: http://gerrit.cloudera.org:8080/8675
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins