docs: Update to more recent jekyll
Since it's not possible for docs generation to work on both macOS and
RHEL 6, let's drop RHEL 6 and move to RHEL 7 as the minimum on Linux to
build the docs.

This also allows us to upgrade the version of yajl-ruby, for which there
is a CVE on versions <= 1.2.2 and 1.3.0.
See https://nvd.nist.gov/vuln/detail/CVE-2017-16516

Change-Id: If3733d1daec263a3ffd1bcd6976d8cd4ba11ddeb
Reviewed-on: http://gerrit.cloudera.org:8080/8678
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Mike Percy <mpercy@apache.org>