consensus: Make queue more debuggable
After the introduction of SafeToEvict() logic in the consensus queue,
the eviction criteria has become more subtle. It would be helpful to get
more information about the state of the leader queue from the web UI to
aid in troubleshooting. This patch dumps the queue information in the
web UI and also includes the time since last communication with each
peer in the per-peer Watermarks table.

Change-Id: I052285a6885c251adbbab62410530fef4769fe72
Reviewed-on: http://gerrit.cloudera.org:8080/8708
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins