web_ui: don't report failed tombstones as live
Previously, the logic to determine the liveness of a tablet in the web
UI was to check that the replica's state message was exactly
"TABLET_DATA_TOMBSTONED". For failed, tombstoned  replicas, this message
is "FAILED(TABLET_DATA_TOMBSTONED)", and as such, these replicas were
being reported as live, failed tablets (scary!).

This patch adjusts the logic to instead search for
"TABLET_DATA_TOMBSTONED" in the message, which should be relatively
short anyway.

I considered actually using the replicas' internal 'state_' members, but
felt that this is more appropriate considering it'd be debatable how to
report STOPPING or FAILED tablets.

A test is added to tablet_server-test that fails without the fix.

Change-Id: I3529c32296b555d59ddde2593360b3b66081ba4b
Reviewed-on: http://gerrit.cloudera.org:8080/8718
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Mike Percy <mpercy@apache.org>
Reviewed-on: http://gerrit.cloudera.org:8080/8739
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins